# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: EnStorage API
  version: 1.0.0
  description: |
    Self-hosted centralized file storage with Google Drive multi-account backend.

    ## Authentication
    Dua mode:
    1. **Sanctum** (web/mobile user) — `Authorization: Bearer <token>`
    2. **API Key** (machine-to-machine) — `X-API-Key: enp_<8>_<40>`

    Endpoint `/api-keys/*` HANYA bisa diakses via Sanctum.

    ## Scopes (API Key only)
    - `read` — GET endpoints
    - `write` — POST/PATCH/PUT (mutation non-destructive)
    - `delete` — DELETE
    - `full` — semua

    ## Rate Limit
    - Sanctum: 60 req/min per user (default Laravel `api` limiter)
    - API Key: 60 req/min per key (`X-RateLimit-Limit`, `X-RateLimit-Remaining`)

    ## Response Envelope
    Sukses: `{ success: true, data, message }` (+ `meta` untuk pagination).
    Gagal: `{ success: false, data: null, message, meta }`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker
  - url: https://{host}/api/v1
    variables:
      host:
        default: storage.example.com
...
paths:
...
  /files:
    get:
      tags: [Files]
      summary: List file dengan filter
      security: [{ sanctum: [] }, { apiKey: [] }]
      parameters:
        - { in: query, name: folder_id, schema: { type: string } }
        - { in: query, name: mime_type, schema: { type: string } }
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: status, schema: { type: string, enum: [pending, uploading, done, failed] } }
        - { in: query, name: sort, schema: { type: string, enum: [name, size, created_at, uploaded_at] } }
        - { in: query, name: dir, schema: { type: string, enum: [asc, desc] } }
        - { in: query, name: per_page, schema: { type: integer, maximum: 100, default: 25 } }
      responses:
        '200': { description: OK }
  /files/{id}:
    get:
      tags: [Files]
      summary: Detail file
      security: [{ sanctum: [] }, { apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Files]
      summary: Rename file
      security: [{ sanctum: [] }, { apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 255 }
      responses:
        '200': { description: OK }
    delete:
      tags: [Files]
      summary: Hapus file (GDrive + record + thumbnail)
      security: [{ sanctum: [] }, { apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      responses:
        '200': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: Resource tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
  schemas:
    Envelope:
      type: object
      properties:
        success: { type: boolean }
        data: { nullable: true }
        message: { type: string }
        meta: { type: object, additionalProperties: true }
...